backend/requirements.dev.txt
Django==5.2.1
django-cors-headers==4.7.0
django-split-settings==1.3.2
python-dotenv==1.1.0
mysqlclient==2.2.7
